Rename `deferral.go` to `deferred.go`
