+++ internal/toproto5/readresource_test.go
@@ -70,7 +70,7 @@ func TestReadResourceResponse(t *testing.T) {
 	}
 
 	testDeferral := &resource.DeferredResponse{
-		Reason: resource.DeferralReasonAbsentPrereq,
+		Reason: resource.DeferredReasonAbsentPrereq,
 	}
 
 	testProto5Deferred := &tfprotov5.Deferred{
